refactor(core): split observability types into types::observability submodule
Continues the `core/src/types.rs` kitchen-sink split. Moves `OtelExporter`,
`ObservabilityConfig`, `IcebergConfig`, and `PartitionGranularity` out of
the flat root namespace into `assistant_core::types::observability::*`,
dropping their entries from the `lib.rs` re-export.

Call-site changes (10 files):
- `crates/runtime/src/telemetry.rs`
- `crates/web-ui/src/lib.rs` (OtelExporter split out of a combined `use`)
- `crates/web-ui/src/backends/iceberg.rs`
- `crates/opentelemetry-exporter-iceberg/src/{lib,catalog,log,metric,partition,span}.rs`
- `crates/opentelemetry-exporter-iceberg/tests/e2e_write_read.rs`

All other use sites within these files were unaffected.

`types.rs` adds `use self::observability::ObservabilityConfig;` so the
`AssistantConfig` field continues to resolve.

Tests:
- `cargo test -p assistant-core --lib` — 183 pass
- `cargo test -p assistant-runtime telemetry` — 7 pass
- `cargo test -p opentelemetry-exporter-iceberg --lib` — 5 pass
- `cargo clippy --workspace -- -D warnings` — green
