+++ crates/opentelemetry-exporter-iceberg/src/catalog.rs
@@ -4,7 +4,7 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use anyhow::{Context, Result};
-use assistant_core::IcebergConfig;
+use assistant_core::types::observability::IcebergConfig;
 use iceberg::io::FileIO;
 use iceberg::memory::{MEMORY_CATALOG_WAREHOUSE, MemoryCatalogBuilder};
 use iceberg::{Catalog, CatalogBuilder, NamespaceIdent};
