+++ crates/opentelemetry-exporter-iceberg/src/span.rs
@@ -5,7 +5,7 @@ use std::sync::{Arc, RwLock};
 use anyhow::Result;
 use arrow_array::{ArrayRef, Int64Array, RecordBatch, StringArray, TimestampMicrosecondArray};
 use arrow_schema::Schema as ArrowSchema;
-use assistant_core::IcebergConfig;
+use assistant_core::types::observability::IcebergConfig;
 use iceberg::spec::DataFileFormat;
 use iceberg::transaction::ApplyTransactionAction;
 use iceberg::transaction::Transaction;
